♻️ Refactor: 모달 시맨틱 태그 및 헤더 아이콘 구조 개선, 컨텍스트 파일명 수정
